Move internal notes tweak to separate file
